+++ bazel/browsers/update-script/README.md
@@ -1,0 +1,4 @@
+## Update script
+
+This is an internal helper script for easing the updates of browsers with `bazel/browsers`. This
+folder is _not_ shipped to the NPM package.
