+++ ng-dev/utils/protos/bazel_test_status_pb.d.ts
