+++ package.json
@@ -34,7 +34,6 @@
     "@bazel/runfiles": "4.6.0",
     "@bazel/terser": "4.6.0",
     "@bazel/typescript": "4.6.0",
-    "@google-cloud/storage": "^5.18.1",
     "@microsoft/api-extractor": "7.19.4",
     "@types/browser-sync": "^2.26.3",
     "@types/node": "16.10.9",
@@ -49,7 +48,6 @@
     "clang-format": "^1.6.0",
     "node-fetch": "^2.6.1",
     "prettier": "^2.3.2",
-    "protobufjs": "^6.11.2",
     "protractor": "^7.0.0",
     "selenium-webdriver": "4.1.1",
     "send": "^0.17.2",
@@ -73,6 +71,7 @@
     "@angular/platform-browser": "14.0.0-next.4",
     "@bazel/bazelisk": "^1.10.1",
     "@bazel/jasmine": "4.6.0",
+    "@google-cloud/storage": "^5.18.1",
     "@octokit/auth-app": "^3.6.0",
     "@octokit/core": "^3.5.1",
     "@octokit/graphql": "^4.8.0",
@@ -116,6 +115,7 @@
     "minimist": "^1.2.5",
     "multimatch": "^5.0.0",
     "nock": "^13.0.3",
+    "protobufjs": "^6.11.2",
     "requirejs": "^2.3.6",
     "rxjs": "^7.4.0",
     "semver": "^7.3.5",
